# backend/openapi.yaml
openapi: 3.1.0
info:
  title: UniversalCart API
  version: 0.5.0
  description: |
    Universal shopping cart Chrome extension backend API.
    Provides cart management, price scouting, checkout orchestration,
    promo code aggregation, deal finding, notifications, and security services.
  contact:
    name: UniversalCart
    url: https://github.com/shroominati/universalcart
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.universalcart.app
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from /api/auth/login or /api/auth/register
    RequestSigning:
      type: apiKey
      in: header
      name: X-UC-Signature
      description: HMAC-SHA256 request signature for extension ↔ backend calls
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        code:
          type: string
          example: VALIDATION_ERROR
        details:
          type: array
          items:
            type: object
            properties:
              location:
                type: string
                enum: [body, params, query]
              field:
                type: string
              message:
                type: string
      required: [error, code, details]
    RateLimitError:
      type: object
      properties:
        error:
          type: string
        retryAfter:
          type: integer
          description: Seconds until rate limit resets
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: integer
          description: Unix timestamp ms
        lastLogin:
          type: integer
    CartItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          maxLength: 500
        price:
          type: number
          minimum: 0
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          default: USD
        image:
          type: string
          format: uri
        sourceUrl:
          type: string
          format: uri
        vendor:
          type: string
          maxLength: 100
        vendorSku:
          type: string
        universalId:
          type: string
        variant:
          type: string
        quantity:
          type: integer
          minimum: 1
          maximum: 999
          default: 1
        status:
          type: string
          enum: [active, watchlist, purchased, removed]
        lowestPrice:
          type: number
        highestPrice:
          type: number
        addedAt:
          type: integer
        updatedAt:
          type: integer
      required: [name, sourceUrl, vendor]
    PriceHistory:
      type: object
      properties:
        price:
          type: number
        recordedAt:
          type: integer
        source:
          type: string
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - BearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
              required: [email, password]
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 1
              required: [email, password]
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      responses:
        "200":
          description: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Cart ────────────────────────────────────────────────────────────────────
...
  /api/cart/sync:
    post:
      tags: [Cart]
      summary: Sync cart items from extension
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/CartItem"
                  minItems: 1
                  maxItems: 500
              required: [items]
      responses:
        "200":
          description: Cart synced
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItem"
      responses:
        "201":
          description: Item added
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/cart/items/{id}:
    patch:
      tags: [Cart]
      summary: Update a cart item
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 0
                  maximum: 999
                status:
                  type: string
                  enum: [active, watchlist, purchased, removed]
                notes:
                  type: string
                  maxLength: 2000
                tags:
                  type: string
                  maxLength: 500
                priceAlertThreshold:
                  type: number
                  minimum: 0
      responses:
        "200":
          description: Item updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Cart]
      summary: Remove item from cart
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Item removed
        "404":
          $ref: "#/components/responses/NotFound"
  /api/cart/items/{id}/watchlist:
    post:
      tags: [Cart]
      summary: Move item to watchlist
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Moved to watchlist
  /api/cart/items/{id}/active:
    post:
      tags: [Cart]
      summary: Move item back to active cart
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Moved to active
  /api/cart/items/{id}/purchased:
    post:
      tags: [Cart]
      summary: Mark item as purchased
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Marked as purchased
  /api/cart/items/{id}/price-history:
    get:
      tags: [Cart]
      summary: Get price history for item
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceHistory"
  /api/cart/items/{id}/analytics:
    get:
      tags: [Cart]
      summary: Get item analytics (price trends, savings)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Item analytics
  /api/cart/items/{id}/refresh:
    post:
      tags: [Cart]
      summary: Force refresh item price
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Price refreshed
  /api/cart/settings:
    patch:
      tags: [Cart]
      summary: Update cart settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                autoRefreshInterval:
                  type: integer
                  minimum: 60000
                  maximum: 86400000
                  description: Auto-refresh interval in milliseconds
                priceDropThreshold:
                  type: number
                  minimum: 0
                  maximum: 100
                  description: Percentage drop to trigger notification
                trackOutOfStock:
                  type: boolean
                maxItems:
                  type: integer
                  minimum: 1
                  maximum: 10000
      responses:
        "200":
          description: Settings updated
        "400":
          $ref: "#/components/responses/BadRequest"

  # ─── Checkout ────────────────────────────────────────────────────────────────
...